Fix size of command preview for Safari
For yet another unknown reason, Safari doesn't agree with the other
browsers and sets the height way too small that it couldn't even fit in
the text. Also when setting overflow to auto it doesn't make enough
space for the scroll bar until literally any CSS option is changed then
it suddenly looks correct. But at least it works fine with scroll.