Assign new colours starting from the end instead of beginning
Instead of creating all colours for the new names and then overwriting
the ones that already had been set, it always starts from the index of
the last colour that has been set. This mainly avoids giving the first
colour to a new name, which is before the previous first one, and
therefore both will have the first colour. So now it gets the next
colour instead.