fix: handle nested arguments in the argument parser
Mostly coming from libraries such as simple-parsing, who nest the
values, but this can also be done manually, so might as well account for
it.
Also this makes sure the values are serialisable, as JSON is very
restrictive and Python doesn't allow to handle it gracefully (as in
having the possibility to define how an object is serialised).